Redid blockchain/slots.rs. Finished blockchain crate reorg.
