+++ blockchain-albatross/src/blockchain/mod.rs
@@ -1,9 +1,12 @@
 pub use blockchain::Blockchain;
+pub use chain_ordering::ChainOrdering;
 
 mod accounts;
 mod blockchain;
+mod chain_ordering;
 mod inherents;
 mod macro_sync;
 mod normal_sync;
+mod slots;
 mod verify;
 mod wrappers;
