+++ blockchain-albatross/src/blockchain/wrappers.rs
@@ -1,5 +1,3 @@
-use std::convert::TryInto;
-
 use parking_lot::{MappedRwLockReadGuard, RwLockReadGuard};
 
 use account::{Account, StakingContract};
@@ -7,17 +5,15 @@ use block::{Block, BlockType, MacroBlock};
 #[cfg(feature = "metrics")]
 use blockchain_base::chain_metrics::BlockchainMetrics;
 use blockchain_base::Direction;
-use database::{ReadTransaction, Transaction, WriteTransaction};
+use database::{Transaction, WriteTransaction};
 use genesis::NetworkInfo;
 use hash::{Blake2bHash, Hash};
 use primitives::policy;
-use primitives::slot::{Slot, Slots, ValidatorSlots};
+use primitives::slot::ValidatorSlots;
 use transaction::Transaction as BlockchainTransaction;
 use utils::merkle;
-use vrf::VrfSeed;
 
 use crate::blockchain_state::BlockchainState;
-use crate::slots::get_slot_at;
 use crate::Blockchain;
 
 /// Wrapper functions
@@ -83,129 +79,6 @@ impl Blockchain {
             Some(chain_info) => include_forks || chain_info.on_main_chain,
             None => false,
         }
-    }
-
-    pub fn get_slots_for_epoch(&self, epoch: u32) -> Option<Slots> {
-        let state = self.state.read();
-        let current_epoch = policy::epoch_at(state.main_chain.head.block_number());
-
-        let slots = if epoch == current_epoch {
-            state.current_slots.as_ref()?.clone()
-        } else if epoch == current_epoch - 1 {
-            state.previous_slots.as_ref()?.clone()
-        } else {
-            let macro_block = self
-                .chain_store
-                .get_block_at(policy::election_block_of(epoch), true, None)?
-                .unwrap_macro();
-            macro_block.try_into().unwrap()
-        };
-
-        Some(slots)
-    }
-
-    pub fn get_validators_for_epoch(&self, epoch: u32) -> Option<ValidatorSlots> {
-        if let Some(slots) = self.get_slots_for_epoch(epoch) {
-            Some(slots.into())
-        } else {
-            None
-        }
-    }
-
-    pub fn next_slots(&self, seed: &VrfSeed, txn_option: Option<&Transaction>) -> Slots {
-        let validator_registry = NetworkInfo::from_network_id(self.network_id)
-            .validator_registry_address()
-            .expect("No ValidatorRegistry");
-
-        let staking_account = self
-            .state
-            .read()
-            .accounts()
-            .get(validator_registry, txn_option);
-
-        if let Account::Staking(ref staking_contract) = staking_account {
-            return staking_contract.select_validators(seed);
-        }
-
-        panic!("Account at validator registry address is not the staking contract!");
-    }
-
-    pub fn next_validators(&self, seed: &VrfSeed, txn: Option<&Transaction>) -> ValidatorSlots {
-        self.next_slots(seed, txn).into()
-    }
-
-    pub fn get_slot_at(
-        &self,
-        block_number: u32,
-        view_number: u32,
-        txn_option: Option<&Transaction>,
-    ) -> Option<(Slot, u16)> {
-        let state = self.state.read_recursive();
-
-        let read_txn;
-        let txn = if let Some(txn) = txn_option {
-            txn
-        } else {
-            read_txn = ReadTransaction::new(&self.env);
-            &read_txn
-        };
-
-        // Gets slots collection from either the cached ones, or from the election block.
-        // Note: We need to handle the case where `state.block_number()` is at a macro block (so `state.current_slots`
-        // was already updated by it, pushing this epoch's slots to `state.previous_slots` and deleting previous'
-        // epoch's slots).
-        let slots_owned;
-        let slots = if policy::epoch_at(state.block_number()) == policy::epoch_at(block_number)
-            && !policy::is_election_block_at(state.block_number())
-        {
-            state.current_slots.as_ref().unwrap_or_else(|| {
-                panic!(
-                    "Missing epoch's slots for block {}.{}",
-                    block_number, view_number
-                )
-            })
-        } else if (policy::epoch_at(state.block_number()) == policy::epoch_at(block_number)
-            && policy::is_election_block_at(state.block_number()))
-            || (policy::epoch_at(state.block_number()) == policy::epoch_at(block_number) + 1
-                && !policy::is_election_block_at(state.block_number()))
-        {
-            state.previous_slots.as_ref().unwrap_or_else(|| {
-                panic!(
-                    "Missing previous epoch's slots for block {}.{}",
-                    block_number, view_number
-                )
-            })
-        } else {
-            let macro_block = self
-                .chain_store
-                .get_block_at(
-                    policy::election_block_before(block_number),
-                    true,
-                    Some(&txn),
-                )?
-                .unwrap_macro();
-
-            // Get slots of epoch
-            slots_owned = macro_block.try_into().unwrap();
-            &slots_owned
-        };
-
-        let prev_info = self
-            .chain_store
-            .get_chain_info_at(block_number - 1, false, Some(&txn))?;
-
-        Some(get_slot_at(block_number, view_number, &prev_info, slots))
-    }
-
-    pub fn get_slot_for_next_block(
-        &self,
-        view_number: u32,
-        txn_option: Option<&Transaction>,
-    ) -> (Slot, u16) {
-        let block_number = self.block_number() + 1;
-
-        self.get_slot_at(block_number, view_number, txn_option)
-            .unwrap()
     }
 
     pub fn get_next_block_type(&self, last_number: Option<u32>) -> BlockType {
